refactor: 헤더 auth action만 클라이언트로 분리
Header 전체가 client component가 되면서
정적 영역까지 클라이언트 번들에 포함되고 있었다.
auth 상태에 의존하는 우측 action 영역만 별도 client component로 분리해
렌더링 영향 범위를 줄였다.

Constraint: 기존 Header UI와 auth 분기 동작은 그대로 유지해야 함
Rejected: Header 전체를 client 유지 | 정적 GNB 영역까지 클라이언트 의존이 불필요하게 퍼짐
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: auth 상태에만 의존하는 인터랙션은 가능한 한 부분 컴포넌트로 분리하고 정적 셸은 server로 유지할 것
Tested: ESLint(Header.tsx/HeaderAuthActions.tsx/header.test.tsx)
Tested: Jest(header.test.tsx)
Tested: TypeScript(noEmit)
Not-tested: 실제 앱에서 Header hydration 타이밍 수동 확인
